factor crosshair code out into its own source file
Fixes #1764
