+++ src/CMakeLists.txt
@@ -27,6 +27,7 @@ set(WOOF_SOURCES
     g_input.c              g_input.h
     hu_command.c           hu_command.h
     hu_coordinates.c       hu_coordinates.h
+    hu_crosshair.c         hu_crosshair.h
     hu_lib.c               hu_lib.h
     hu_obituary.c          hu_obituary.h
     hu_stuff.c             hu_stuff.h
