+++ src/hu_stuff.h
@@ -88,8 +88,6 @@ enum
 extern int hud_type;
 extern boolean draw_crispy_hud;
 
-extern int hud_crosshair;
-
 enum
 {
   HUD_WIDGET_OFF,
@@ -99,27 +97,9 @@ enum
   HUD_WIDGET_ADVANCED,
 };
 
-typedef enum
-{
-  crosstarget_off,
-  crosstarget_highlight,
-  crosstarget_health, // [Alaux] Color crosshair by target health
-} crosstarget_t;
-extern crosstarget_t hud_crosshair_target;
-
-// [Alaux] Lock crosshair on target
-extern boolean hud_crosshair_lockon;
-extern struct mobj_s *crosshair_target;
-void HU_UpdateCrosshairLock(int x, int y);
-void HU_DrawCrosshair(void);
-
-extern int hud_crosshair_color;
-
-#define HU_CROSSHAIRS 10
-extern const char *crosshair_lumps[HU_CROSSHAIRS];
-extern const char *crosshair_strings[HU_CROSSHAIRS];
-
 void HU_BindHUDVariables(void);
+
+byte* ColorByHealth(int health, int maxhealth, boolean invul);
 
 #endif
